v.o.1 Stock: edit & delete buttons
Made it so that when user clicks on the delete button of the item they want to remove it will remove it. And when they click on the edit button a modal to edit the item will appear with the fields populatied witht the clicked item. Still need to work on feautre where it will update that specific item and not create a new item.